# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  error_code.ErrorCodeConst:
    enum:
    - CannotDeleteLastSSOBinding
    - DirectoryNotFound
    - FileAlreadyExists
    - FileNotFound
    - FileOperationFailed
    - FileTooLarge
    - Forbidden
    - InternalServerError
    - InvalidAccessToken
    - InvalidCredentials
    - InvalidFilePath
    - InvalidFileType
    - InvalidParameter
    - InvalidParameters
    - InvalidRecoveryCode
    - InvalidRefreshToken
    - InvalidTotpCode
    - OauthTokenUnavailable
    - PasswordLoginIsNotEnabled
    - SSOProviderAccountAlreadyBinded
    - StorageQuotaExceeded
    - TokenNotFound
    - TwoFaAlreadyEnabled
    - TwoFaTotpIsRequiredForLogin
    - Unauthorized
    - UserAlreadyExists
    - UserNotFound
    type: string
    x-enum-varnames:
    - ErrorCodeCannotDeleteLastSSOBinding
    - ErrorCodeDirectoryNotFound
    - ErrorCodeFileAlreadyExists
    - ErrorCodeFileNotFound
    - ErrorCodeFileOperationFailed
    - ErrorCodeFileTooLarge
    - ErrorCodeForbidden
    - ErrorCodeInternalServerError
    - ErrorCodeInvalidAccessToken
    - ErrorCodeInvalidCredentials
    - ErrorCodeInvalidFilePath
    - ErrorCodeInvalidFileType
    - ErrorCodeInvalidParameter
    - ErrorCodeInvalidParameters
    - ErrorCodeInvalidRecoveryCode
    - ErrorCodeInvalidRefreshToken
    - ErrorCodeInvalidTotpCode
    - ErrorCodeOauthTokenUnavailable
    - ErrorCodePasswordLoginIsNotEnabled
    - ErrorCodeSSOProviderAccountAlreadyBinded
    - ErrorCodeStorageQuotaExceeded
    - ErrorCodeTokenNotFound
    - ErrorCodeTwoFaAlreadyEnabled
    - ErrorCodeTwoFaTotpIsRequiredForLogin
    - ErrorCodeUnauthorized
    - ErrorCodeUserAlreadyExists
    - ErrorCodeUserNotFound
...
  swagger.BaseFailResponse:
    properties:
      error_code:
        $ref: '#/definitions/error_code.ErrorCodeConst'
      message:
        example: Internal server error
        type: string
      request_id:
        example: 3c55e76b-21b0-41e9-9780-206283053886
        type: string
      status:
        example: error
        type: string
    required:
    - error_code
    - message
    - request_id
    - status
    type: object
...
  swagger.BaseSuccessResponse-user_CreateUserResponseDto:
    properties:
      data:
        $ref: '#/definitions/user.CreateUserResponseDto'
      message:
        example: ""
        type: string
      request_id:
        example: 3c55e76b-21b0-41e9-9780-206283053886
        type: string
      status:
        example: "1"
        type: string
    required:
    - data
    - message
    - request_id
    - status
    type: object
...
  user.CreateUserRequestDto:
    properties:
      password:
        example: password
        maxLength: 32
        type: string
      username:
        example: username
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  user.CreateUserResponseDto:
    type: object
...
host: localhost:8080
info:
  contact:
    name: WonderfulSoap
  description: This is a sample server
  termsOfService: http://swagger.io/terms/
  title: My-Golang-Framework
  version: "1.0"
paths:
...
  /api/v1/user/create:
    post:
      consumes:
      - application/json
      description: Create a new user with username and password
      parameters:
      - description: New user information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.BaseSuccessResponse-user_CreateUserResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.BaseFailResponse'
      summary: Create user
      tags:
      - User
...
swagger: "2.0"